astxml2wiki.xslt: Add support for <example> tags
This patch parses out <example> tags and converts them into code macro blocks.
It sets linenumbers on by default, and looks for the presence of the title
attribute to set the title, if available. The language attribute is treated
in the same fashion as the title attribute.

With this patch, code/dialplan examples can be placed in the Asterisk source
and formatted as code blocks on wiki.asterisk.org.