publish-docs: Update publish script to work with Jenkins for rest api docs
The publish script was looking for a 'rest-api' directory beneath 'doc/'. If it
existed it [re]built the rest api documentation and published it. However, if
this directory did not exist then no rest api documentation would be built,
updated and published.

This patch now checks for the 'rest-api' directory directly under the top level
Asterisk directory instead.

Also there were a couple of checks that used 'svn'. Updated these checks to now
use 'git'.

Change-Id: Ia5b71adb33327f6405192df62a42a99e70fb812e