astxml2wiki.xslt: Improve formatting of arguments outside of syntax sections
This patch causes the parameters of an argument to be displayed with the
argument, if it has them. Required parameters are bolded, and - if a
custom separator is used - the separator is now shown for the
parameters. Previously this information was not outputted on the wiki
documentation.

Additionally, a newline is now emitted after an enumlist, which helps to
properly unindent content that follows an enumlist.

Change-Id: Iebb8f8f1d588c1cb39fe7a5bf6491b86a072fbbc