astxml2wiki: Format <info> blocks better; include <example> blocks
This patch improves the formatting of <info> blocks in a few ways.
First, prior to this patch, it was always assumed that an <info> block
existed within an enumlist. The bulleting would continue at that level.
We now only continue the bulleting if a bullet was passed in, and, if a
bullet was passed in, we start the "Technology" title at the current
level as opposed to the next level. This slightly improves the bulleted
formatting by putting Technology specific items at the same bullet level
as the items they represent.

Since <info> blocks are now not always bulleted (unless a bullet is
passed in), we can now display <example> blocks in them as well. The
appropriate templates are now called from the <info> template.

Finally, this patch removes the newline that was committed in the
previous patch for enumlists, as it wasn't actually necessary (and
negatively impacted enumlists called from inside an enum tag).

Change-Id: I2f4a93c46dd6fa49abc3d3c4d473e4052de018ac