publish-docs: Stop needlessly updating wiki pages.
This set of changes tweaks the generated wiki pages to match
that of Confluence. There were some cases where Confluence
would insert, remove, or reorder things causing our generated
copy to differ from the real copy and thus forcing an update
of the wiki page each run which did nothing.

The see-also ref nodes for some things were getting double
modified with link information due to the same ref node being
shared. We now duplicate the full node to ensure that we
don't modify any of the ref nodes multiple times.

When escaping the '_' character using '\\' Confluence will
insert a '<br class="atl-forced-newline" />' in some cases but
when updated on the wiki it is removed. This is now also removed
from our generated copy.

The generated copy now replaces '&#35;' with its real value
of '#' which matches the Confluence wiki page.

Some configuration XML does not include a synopsis. This would
cause us to place an empty h1 in our generated copy. Confluence
would remove this (as it was meaningless to have present). We
now only add this if a synopsis is present.

The example template has been reordered slightly to ensure that
the linenumbers value is present after the title and language.
Despite providing linenumbers as the first thing to Confluence
it reordered it internally.

The ARI process has now been updated to replace the '&quot;' we
receive from Confluence with '"' which matches our generated
copy.

The unchanged value now works and reflects how many pages
were unchanged.

Change-Id: I86add2a6edadb2ed1bf5869025623d9624a1a4f1